Split file into smaller parts
