+++ hacker_news_bot/main.py
@@ -1,9 +1,9 @@
 import os
 from collections.abc import Iterable
 
-from configure import setup_logger
-from datasources.hacker_news import HackerNews, Item
-from toot_formatter import prepare_toot_message
+from hacker_news_bot.configure import setup_logger
+from hacker_news_bot.datasources.hacker_news import HackerNews, Item
+from hacker_news_bot.translator import prepare_toot_message
 
 setup_logger(os.environ.get("LOGLEVEL", "ERROR").upper())
